Allow directory client to be build only with some clients. e,g, reader

+++ src/main/java/com/aserto/directory/v3/DirectoryClient.java
@@ -50,22 +50,48 @@ public class DirectoryClient implements DirectoryClientReader,
     private ModelGrpc.ModelBlockingStub modelClient;
     private ModelGrpc.ModelStub modelClientAsync;
 
-    public DirectoryClient(ManagedChannel channelBuilder) {
-        DirectoryClientBuilder dirClientBuilder = new DirectoryClientBuilder(channelBuilder);
-        readerClient = dirClientBuilder.getReaderClient();
-        writerClient = dirClientBuilder.getWriterClient();
-        importerClient = dirClientBuilder.getImporterClient();
-        exporterClient = dirClientBuilder.getExporterClient();
-        modelClient = dirClientBuilder.getModelClient();
-        modelClientAsync = dirClientBuilder.getModelClientAsync();
+    public DirectoryClient(ManagedChannel readerChannel, ManagedChannel writerChannel, ManagedChannel importerChannel, ManagedChannel exporterChannel, ManagedChannel modelChannel) {
+        if (readerChannel != null) {
+            readerClient = ReaderGrpc.newBlockingStub(readerChannel);
+        }
+
+        if (writerChannel != null) {
+            writerClient = WriterGrpc.newBlockingStub(writerChannel);
+        }
+
+        if (importerChannel != null) {
+            importerClient = ImporterGrpc.newStub(importerChannel);
+        }
+
+        if (exporterChannel != null) {
+            exporterClient = ExporterGrpc.newBlockingStub(exporterChannel);
+        }
+
+        if (modelChannel != null) {
+            modelClient = ModelGrpc.newBlockingStub(modelChannel);
+            modelClientAsync = ModelGrpc.newStub(modelChannel);
+        }
+    }
+
+    public DirectoryClient(ManagedChannel managedChannel) {
+        readerClient = ReaderGrpc.newBlockingStub(managedChannel);
+        writerClient = WriterGrpc.newBlockingStub(managedChannel);
+        importerClient = ImporterGrpc.newStub(managedChannel);
+        exporterClient = ExporterGrpc.newBlockingStub(managedChannel);
+        modelClient = ModelGrpc.newBlockingStub(managedChannel);
+        modelClientAsync = ModelGrpc.newStub(managedChannel);
     }
 
     @Override
-    public GetObjectResponse getObject(String type, String id) {
+    public GetObjectResponse getObject(String type, String id) throws UninitilizedClientException {
         return getObject(type, id, false);
     }
     @Override
-    public GetObjectResponse getObject(String type, String id, boolean withRelations) {
+    public GetObjectResponse getObject(String type, String id, boolean withRelations) throws UninitilizedClientException {
+        if (readerClient == null) {
+            throw new UninitilizedClientException("Reader client is not initialized");
+        }
+
         return readerClient.getObject(GetObjectRequest.newBuilder()
                 .setObjectType(type)
                 .setObjectId(id)
@@ -74,12 +100,16 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public GetObjectsResponse getObjects(String type) {
+    public GetObjectsResponse getObjects(String type) throws UninitilizedClientException {
         return getObjects(type, 100, "");
     }
 
     @Override
-    public GetObjectsResponse getObjects(String type, int pageSize, String pageToken) {
+    public GetObjectsResponse getObjects(String type, int pageSize, String pageToken) throws UninitilizedClientException {
+        if (readerClient == null) {
+            throw new UninitilizedClientException("Reader client is not initialized");
+        }
+
         return readerClient.getObjects(GetObjectsRequest.newBuilder()
                 .setObjectType(type)
                 .setPage(buildPaginationRequest(pageSize, pageToken))
@@ -87,7 +117,11 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public GetObjectManyResponse getObjectManyRequest(List<ObjectIdentifier> objectIdentifiers) {
+    public GetObjectManyResponse getObjectManyRequest(List<ObjectIdentifier> objectIdentifiers) throws UninitilizedClientException {
+        if (readerClient == null) {
+            throw new UninitilizedClientException("Reader client is not initialized");
+        }
+
         return readerClient.getObjectMany(GetObjectManyRequest.newBuilder()
                 .addAllParam(new ObjectIdentifierList(objectIdentifiers))
                 .build());
@@ -101,17 +135,21 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public GetRelationResponse getRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId) {
+    public GetRelationResponse getRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId) throws UninitilizedClientException {
         return getRelation(objectType, objectId, relationName, subjectType, subjectId, "", false);
     }
 
     @Override
-    public GetRelationResponse getRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, String subjectRelation) {
+    public GetRelationResponse getRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, String subjectRelation) throws UninitilizedClientException {
         return getRelation(objectType, objectId, relationName, subjectType, subjectId, subjectRelation, false);
     }
 
     @Override
-    public GetRelationResponse getRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, String subjectRelation, boolean withObjects) {
+    public GetRelationResponse getRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, String subjectRelation, boolean withObjects) throws UninitilizedClientException {
+        if (readerClient == null) {
+            throw new UninitilizedClientException("Reader client is not initialized");
+        }
+
         return readerClient.getRelation(GetRelationRequest.newBuilder()
                 .setObjectType(objectType)
                 .setObjectId(objectId)
@@ -124,17 +162,25 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public GetRelationsResponse getRelations(GetRelationsRequest relationsRequest) {
+    public GetRelationsResponse getRelations(GetRelationsRequest relationsRequest) throws UninitilizedClientException {
+        if (readerClient == null) {
+            throw new UninitilizedClientException("Reader client is not initialized");
+        }
+
         return readerClient.getRelations(relationsRequest);
     }
 
     @Override
-    public CheckPermissionResponse checkPermission(String objectType, String objectId, String subjectType, String subjectId, String permissionName) {
+    public CheckPermissionResponse checkPermission(String objectType, String objectId, String subjectType, String subjectId, String permissionName) throws UninitilizedClientException {
         return checkPermission(objectType, objectId, subjectType, subjectId, permissionName, false);
     }
 
     @Override
-    public CheckPermissionResponse checkPermission(String objectType, String objectId, String subjectType, String subjectId, String permissionName, boolean trace) {
+    public CheckPermissionResponse checkPermission(String objectType, String objectId, String subjectType, String subjectId, String permissionName, boolean trace) throws UninitilizedClientException {
+        if (readerClient == null) {
+            throw new UninitilizedClientException("Reader client is not initialized");
+        }
+
         return readerClient.checkPermission(CheckPermissionRequest.newBuilder()
                 .setObjectType(objectType)
                 .setObjectId(objectId)
@@ -146,12 +192,16 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public CheckRelationResponse checkRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId) {
+    public CheckRelationResponse checkRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId) throws UninitilizedClientException {
         return checkRelation(objectType, objectId, relationName, subjectType, subjectId, false);
     }
 
     @Override
-    public CheckRelationResponse checkRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, boolean trace) {
+    public CheckRelationResponse checkRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, boolean trace) throws UninitilizedClientException {
+        if (readerClient == null) {
+            throw new UninitilizedClientException("Reader client is not initialized");
+        }
+
         return readerClient.checkRelation(CheckRelationRequest.newBuilder()
                 .setObjectType(objectType)
                 .setObjectId(objectId)
@@ -163,12 +213,16 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public CheckResponse check(String objectType, String objectId, String relationName, String subjectType, String subjectId) {
+    public CheckResponse check(String objectType, String objectId, String relationName, String subjectType, String subjectId) throws UninitilizedClientException {
         return check(objectType, objectId, relationName, subjectType, subjectId, false);
     }
 
     @Override
-    public CheckResponse check(String objectType, String objectId, String relationName, String subjectType, String subjectId, boolean trace) {
+    public CheckResponse check(String objectType, String objectId, String relationName, String subjectType, String subjectId, boolean trace) throws UninitilizedClientException {
+        if (readerClient == null) {
+            throw new UninitilizedClientException("Reader client is not initialized");
+        }
+
         return readerClient.check(CheckRequest.newBuilder()
                 .setObjectType(objectType)
                 .setObjectId(objectId)
@@ -185,12 +239,16 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public SetObjectResponse setObject(String type, String id) {
+    public SetObjectResponse setObject(String type, String id) throws UninitilizedClientException {
         return setObject(type, id, "", Struct.newBuilder().build(), "");
     }
 
     @Override
-    public SetObjectResponse setObject(String type, String id, String displayName, Struct properties, String hash) {
+    public SetObjectResponse setObject(String type, String id, String displayName, Struct properties, String hash) throws UninitilizedClientException {
+        if (writerClient == null) {
+            throw new UninitilizedClientException("Writer client is not initialized");
+        }
+
         Instant time = Instant.now();
         Timestamp timestamp = Timestamp.newBuilder().setSeconds(time.getEpochSecond())
                 .setNanos(time.getNano()).build();
@@ -209,12 +267,16 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public DeleteObjectResponse deleteObject(String type, String id) {
+    public DeleteObjectResponse deleteObject(String type, String id) throws UninitilizedClientException {
         return deleteObject(type, id, false);
     }
 
     @Override
-    public DeleteObjectResponse deleteObject(String type, String id, boolean withRelations) {
+    public DeleteObjectResponse deleteObject(String type, String id, boolean withRelations) throws UninitilizedClientException {
+        if (writerClient == null) {
+            throw new UninitilizedClientException("Writer client is not initialized");
+        }
+
         return writerClient.deleteObject(DeleteObjectRequest.newBuilder()
                 .setObjectType(type)
                 .setObjectId(id)
@@ -223,7 +285,11 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public SetRelationResponse setRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId) {
+    public SetRelationResponse setRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId) throws UninitilizedClientException {
+        if (writerClient == null) {
+            throw new UninitilizedClientException("Writer client is not initialized");
+        }
+
         Relation relation = Relation.newBuilder()
                 .setObjectType(objectType)
                 .setObjectId(objectId)
@@ -236,7 +302,11 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public SetRelationResponse setRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, String subjectRelation) {
+    public SetRelationResponse setRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, String subjectRelation) throws UninitilizedClientException {
+        if (writerClient == null) {
+            throw new UninitilizedClientException("Writer client is not initialized");
+        }
+
         Relation relation = Relation.newBuilder()
                 .setObjectType(objectType)
                 .setObjectId(objectId)
@@ -250,7 +320,11 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public SetRelationResponse setRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, String subjectRelation, String hash) {
+    public SetRelationResponse setRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, String subjectRelation, String hash) throws UninitilizedClientException {
+        if (writerClient == null) {
+            throw new UninitilizedClientException("Writer client is not initialized");
+        }
+
         Relation relation = Relation.newBuilder()
                 .setObjectType(objectType)
                 .setObjectId(objectId)
@@ -265,7 +339,11 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public DeleteRelationResponse deleteRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId) {
+    public DeleteRelationResponse deleteRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId) throws UninitilizedClientException {
+        if (writerClient == null) {
+            throw new UninitilizedClientException("Writer client is not initialized");
+        }
+
         return writerClient.deleteRelation(DeleteRelationRequest.newBuilder()
                 .setObjectType(objectType)
                 .setObjectId(objectId)
@@ -276,7 +354,11 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public DeleteRelationResponse deleteRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, String subjectRelation) {
+    public DeleteRelationResponse deleteRelation(String objectType, String objectId, String relationName, String subjectType, String subjectId, String subjectRelation) throws UninitilizedClientException {
+        if (writerClient == null) {
+            throw new UninitilizedClientException("Writer client is not initialized");
+        }
+
         return writerClient.deleteRelation(DeleteRelationRequest.newBuilder()
                 .setObjectType(objectType)
                 .setObjectId(objectId)
@@ -288,7 +370,11 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public GetManifestResponse getManifest() {
+    public GetManifestResponse getManifest() throws UninitilizedClientException {
+        if (modelClient == null) {
+            throw new UninitilizedClientException("Model client is not initialized");
+        }
+
         GetManifestRequest manifestRequest = GetManifestRequest.newBuilder().build();
         Iterator<GetManifestResponse> manifestResponses =  modelClient.getManifest(manifestRequest);
 
@@ -317,7 +403,11 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public void setManifest(String manifest) throws InterruptedException {
+    public void setManifest(String manifest) throws InterruptedException, UninitilizedClientException {
+        if (modelClientAsync == null) {
+            throw new UninitilizedClientException("Model client is not initialized");
+        }
+
         CountDownLatch latch = new CountDownLatch(1);
 
         StreamObserver<SetManifestResponse> readStream = new StreamObserver<SetManifestResponse>() {
@@ -357,12 +447,20 @@ public class DirectoryClient implements DirectoryClientReader,
 
 
     @Override
-    public DeleteManifestResponse deleteManifest() {
+    public DeleteManifestResponse deleteManifest() throws UninitilizedClientException {
+        if (modelClient == null) {
+            throw new UninitilizedClientException("Model client is not initialized");
+        }
+
         return modelClient.deleteManifest(DeleteManifestRequest.newBuilder().build());
     }
 
     @Override
-    public void importData(Stream<ImportElement> importStream) throws InterruptedException {
+    public void importData(Stream<ImportElement> importStream) throws InterruptedException, UninitilizedClientException {
+        if (importerClient == null) {
+            throw new UninitilizedClientException("Import client is not initialized");
+        }
+
         CountDownLatch latch = new CountDownLatch(1);
         StreamObserver<ImportResponse> readStream = new StreamObserver<ImportResponse>() {
             @Override
@@ -400,14 +498,22 @@ public class DirectoryClient implements DirectoryClientReader,
     }
 
     @Override
-    public Iterator<ExportResponse> exportData(Option options) {
+    public Iterator<ExportResponse> exportData(Option options) throws UninitilizedClientException {
+        if (exporterClient == null) {
+            throw new UninitilizedClientException("Export client is not initialized");
+        }
+
         return exporterClient.export(ExportRequest.newBuilder()
                 .setOptions(options.getNumber())
                 .build());
     }
 
     @Override
-    public Iterator<ExportResponse> exportData(Option options, Timestamp startFrom) {
+    public Iterator<ExportResponse> exportData(Option options, Timestamp startFrom) throws UninitilizedClientException {
+        if (exporterClient == null) {
+            throw new UninitilizedClientException("Export client is not initialized");
+        }
+
         return exporterClient.export(ExportRequest.newBuilder()
                 .setOptions(options.getNumber())
                 .setStartFrom(startFrom)
